requirements.txt
flask
flask-smorest
# python-dotenv
sqlalchemy
flask-sqlalchemy
flask-jwt-extended
flask-migrate
gunicorn
flask-bcrypt
marshmallow_enum
psycopg2-binary
requests
boto3
opencv-python-headless
google-cloud-bigquery
gcloud